requirements.txt
Flask
azure-ai-textanalytics
